Implemented OpenStreetMaps on locations - Improved locations such as streamlining validation - Schema for OSM fields (id and type) implemented - Various schema cleanup - Implemented map and address lookup from wayfinderweb (old) to current site. - Fixed an issue where the navbar color is not compliant to brand guidelines of NMTAFE

+++ database/migrations/2020_07_18_100144_create_events_table.php
@@ -19,8 +19,8 @@ class CreateEventsTable extends Migration
             $table->string('description', 255);
             $table->datetime('start');
             $table->datetime('end');
-            $table->unsignedInteger('location_id');
-            $table->unsignedInteger('upload_id')->nullable()->references('id')->on('uploads');
+            $table->unsignedBigInteger('location_id');
+            $table->unsignedBigInteger('upload_id')->nullable()->references('id')->on('uploads');
             $table->timestamp("created_at")->default(DB::raw("CURRENT_TIMESTAMP"));
             $table->timestamp("updated_at")->default(DB::raw("CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP"));
         });
